requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.2.0
async-timeout==4.0.3
attrs==23.2.0
bcrypt==4.0.1
bidict==0.22.1
blinker==1.7.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
dataclasses-json==0.6.4
distlib==0.3.8
distro==1.9.0
exceptiongroup==1.2.0
filelock==3.13.1
Flask==3.0.2
Flask-Cors==4.0.0
Flask-Mail==0.9.1
flask-marshmallow==1.2.0
Flask-SocketIO==5.3.6
Flask-SQLAlchemy==3.1.1
frozenlist==1.4.1
h11==0.14.0
httpcore==1.0.2
httpx==0.26.0
idna==3.6
importlib-metadata==7.0.1
itsdangerous==2.1.2
Jinja2==3.1.3
jsonpatch==1.33
jsonpointer==2.4
langchain==0.0.354
langchain-community==0.0.19
langchain-core==0.1.22
langchain-openai==0.0.5
langsmith==0.0.87
MarkupSafe==2.1.3
marshmallow==3.20.2
marshmallow-sqlalchemy==1.0.0
multidict==6.0.5
mypy-extensions==1.0.0
numpy==1.26.1
openai==1.12.0
packaging==23.2
platformdirs==4.1.0
protobuf==4.24.4
psycopg2-binary==2.9.9
pydantic==2.6.1
pydantic_core==2.16.2
PyJWT==2.8.0
python-dotenv==1.0.0
python-engineio==4.8.2
python-socketio==5.11.0
PyYAML==6.0.1
regex==2023.12.25
requests==2.31.0
simple-websocket==1.0.0
sniffio==1.3.0
SQLAlchemy==2.0.26
tenacity==8.2.3
tiktoken==0.5.2
tqdm==4.66.2
typing-inspect==0.9.0
typing_extensions==4.9.0
urllib3==2.2.0
virtualenv==20.25.0
Werkzeug==3.0.1
wsproto==1.2.0
yarl==1.9.4
zipp==3.17.0
